+++ app/models.py
@@ -29,7 +29,7 @@ class Product(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(255), nullable=False)
-    price = db.Column(db.Float, nullable=False)
+    price = db.Column(db.Integer, nullable=False)
     amount = db.Column(db.Integer, nullable=False)
     description = db.Column(db.Text)
     image = db.Column(db.String(255))
